refactor(identity): lower pane and channel classifiers
Put sidebar chrome beside pane identity and channel fallback beside IDs.
