+++ refactor-target.toml
@@ -158,7 +158,7 @@ surface-budget = 592
 
 [[module]]
 path = "crates/rimz/src/ids.rs"
-surface-budget = 38
+surface-budget = 39
 
 [[module]]
 path = "crates/rimz/src/lane.rs"
@@ -212,7 +212,7 @@ surface-budget = 4
 [[module]]
 path = "crates/rimz/src/pane"
 upward-dependencies = ["store::snapshot"]
-surface-budget = 32
+surface-budget = 33
 
 [[module]]
 path = "crates/rimz/src/proc"
